+++ _pages/api/hello.ts
