+++ _pages/index.tsx
@@ -1,6 +1,6 @@
 import Head from "next/head";
 import styled from "styled-components";
-import { size, device, truncate } from "@/lib/style";
+import { size, device, truncate } from "@/styles/style";
 import Link from "next/link";
 
 const Frame = styled.div`
