+++ _pages/sign.tsx
@@ -1,6 +1,6 @@
 import Head from "next/head";
 import styled from "styled-components";
-import { size, device, truncate } from '@/lib/style'
+import { size, device, truncate } from "@/styles/style";
 import Input from "@/components/Input";
 import { useEffect, useState } from "react";
 import { ChzzkClient } from "chzzk";
@@ -23,7 +23,7 @@ const Frame = styled.div`
   @media ${device.mobile} {
     padding: 80px 10px;
   }
-`
+`;
 
 const Text1 = styled.p`
   text-align: center;
@@ -33,7 +33,7 @@ const Text1 = styled.p`
   @media ${device.mobile} {
     font: 600 12px/1.5 var(--font-default);
   }
-`
+`;
 
 const Text2 = styled.p`
   text-align: center;
@@ -43,13 +43,13 @@ const Text2 = styled.p`
   @media ${device.mobile} {
     font: 600 18px/1.5 var(--font-default);
   }
-`
+`;
 
 const TopText = styled.div`
   display: flex;
   flex-direction: column;
   gap: 24px;
-`
+`;
 
 const Inputs = styled.div`
   display: flex;
@@ -57,11 +57,11 @@ const Inputs = styled.div`
   gap: 10px;
   width: 100%;
   max-width: 800px;
-`
+`;
 
 const InputChannel = styled(Input)`
   flex: 1;
-`
+`;
 
 const BtnSmall = styled.a`
   display: flex;
@@ -84,7 +84,7 @@ const BtnSmall = styled.a`
     height: 48px;
     width: 80px;
   }
-`
+`;
 
 function isValidText(text: string): boolean {
   // 정규 표현식을 사용하여 32자의 영어 알파벳(대소문자 구분 없음)과 숫자로만 이루어져 있는지 확인
@@ -95,88 +95,88 @@ function isValidText(text: string): boolean {
 function getLastPathSegment(input: string): string {
   try {
     const url = new URL(input);
-    const pathSegments = url.pathname.split('/');
+    const pathSegments = url.pathname.split("/");
     return pathSegments.filter(Boolean).pop() || input;
   } catch {
     return input;
   }
 }
 
-
 export default function Home() {
+  const [channelId, setChannelId] = useState<string>("");
+  const [find, setFind] = useState<ChannelType | null>(null);
+  const { channel, setChannel, setVoice } = useGlobalOptionStore();
 
-  const [ channelId, setChannelId ] = useState<string>('')
-  const [ find, setFind ] = useState<ChannelType | null>(null)
-  const { channel, setChannel, setVoice } = useGlobalOptionStore()
-
-  const router = useRouter()
+  const router = useRouter();
 
   const onSubmit = async () => {
-    const getLastPath = getLastPathSegment(channelId)
-    const isValidUrl = isValidText(getLastPath)
-    
-    if(!isValidUrl){
-      alert('유효하지 않은 주소(채널ID)입니다.\n다시 한 번 확인해주세요.')
-      return
+    const getLastPath = getLastPathSegment(channelId);
+    const isValidUrl = isValidText(getLastPath);
+
+    if (!isValidUrl) {
+      alert("유효하지 않은 주소(채널ID)입니다.\n다시 한 번 확인해주세요.");
+      return;
     }
 
     const options = {
       baseUrls: {
-        chzzkBaseUrl: '/api/proxy/chzzkBase',
-        gameBaseUrl: '/api/proxy/gameBase'
-      }
-    }
+        chzzkBaseUrl: "/api/proxy/chzzkBase",
+        gameBaseUrl: "/api/proxy/gameBase",
+      },
+    };
 
-    const client = new ChzzkClient(options)
+    const client = new ChzzkClient(options);
 
-    const findChannel = await client.channel(getLastPath)
+    const findChannel = await client.channel(getLastPath);
     //console.log(findChannel)
 
-    if(findChannel === null){
-      alert('유효하지 않은 주소(채널ID)입니다.\n다시 한 번 확인해주세요.')
-      return
+    if (findChannel === null) {
+      alert("유효하지 않은 주소(채널ID)입니다.\n다시 한 번 확인해주세요.");
+      return;
     }
 
     const channelData = {
       channelId: findChannel.channelId,
-      channelImageUrl: findChannel.channelImageUrl ? findChannel.channelImageUrl : '',
+      channelImageUrl: findChannel.channelImageUrl
+        ? findChannel.channelImageUrl
+        : "",
       channelName: findChannel.channelName,
       verifiedMark: findChannel.verifiedMark,
-      followerCount: findChannel.followerCount
-    }
+      followerCount: findChannel.followerCount,
+    };
 
-    setFind(channelData)
-  }
+    setFind(channelData);
+  };
 
   const onEnter = (event: React.KeyboardEvent<HTMLInputElement>) => {
-    if(event.key === 'Enter'){
-      onSubmit()
+    if (event.key === "Enter") {
+      onSubmit();
     }
-  }
+  };
 
   const onSetChannel = async () => {
-    if(!find){
-      alert('잘못된 요청입니다!')
-      return
+    if (!find) {
+      alert("잘못된 요청입니다!");
+      return;
     }
 
-    const voices = await getVoices()
-    const voicesKR = voices.filter(voice => voice.lang === 'ko-KR')
-    const findGoogleVoice = voicesKR.find(e => e.name.startsWith('Google'))
+    const voices = await getVoices();
+    const voicesKR = voices.filter((voice) => voice.lang === "ko-KR");
+    const findGoogleVoice = voicesKR.find((e) => e.name.startsWith("Google"));
 
-    if(findGoogleVoice) setVoice(findGoogleVoice.name)
-    else if(voicesKR.length !== 0) setVoice(voicesKR[0].name)
+    if (findGoogleVoice) setVoice(findGoogleVoice.name);
+    else if (voicesKR.length !== 0) setVoice(voicesKR[0].name);
 
-    setChannel(find)
-    router.push('/')
-  }
+    setChannel(find);
+    router.push("/");
+  };
 
   useEffect(() => {
-    if(channel.channelId !== ''){
-      router.push('/')
+    if (channel.channelId !== "") {
+      router.push("/");
     }
-  }, [channel])
-  
+  }, [channel]);
+
   return (
     <>
       <Head>
@@ -187,24 +187,42 @@ export default function Home() {
       </Head>
       <Frame>
         <TopText>
-          <Text1>안녕하세요!<br/>처음 오셨나요?</Text1>
-          <Text2>추첨 및 투표를 진행할 치지직 채널의 주소(URL)를 알려주세요!</Text2>
+          <Text1>
+            안녕하세요!
+            <br />
+            처음 오셨나요?
+          </Text1>
+          <Text2>
+            추첨 및 투표를 진행할 치지직 채널의 주소(URL)를 알려주세요!
+          </Text2>
         </TopText>
         <Inputs>
           <InputChannel
             type="text"
             placeholder="ex) https://chzzk.naver.com/ca1850b2eceb7f86146695fd9bb9cefc"
             value={channelId}
-            onChange={(event: React.ChangeEvent<HTMLInputElement>) => setChannelId(event.target.value)}
+            onChange={(event: React.ChangeEvent<HTMLInputElement>) =>
+              setChannelId(event.target.value)
+            }
             onKeyUp={onEnter}
           />
           <BtnSmall onClick={onSubmit}>등록</BtnSmall>
         </Inputs>
-        <Text1>채널ID를 직접 입력하셔도 괜찮아요!<br/>등록된 채널ID는 이후 설정 메뉴에서 변경 가능합니다<br/>데이터는 컴퓨터에 저장되며, 서버에는 전송되지 않습니다</Text1>
+        <Text1>
+          채널ID를 직접 입력하셔도 괜찮아요!
+          <br />
+          등록된 채널ID는 이후 설정 메뉴에서 변경 가능합니다
+          <br />
+          데이터는 컴퓨터에 저장되며, 서버에는 전송되지 않습니다
+        </Text1>
       </Frame>
-      {
-        find !== null && <FindChannel find={find} onCancel={() => setFind(null)} onAccept={onSetChannel} />
-      }
+      {find !== null && (
+        <FindChannel
+          find={find}
+          onCancel={() => setFind(null)}
+          onAccept={onSetChannel}
+        />
+      )}
     </>
-  )
+  );
 }
